+++ llvm/bindings/go/llvm/executionengine_test.go
@@ -81,7 +81,6 @@ func TestFactorial(t *testing.T) {
 	pass := NewPassManager()
 	defer pass.Dispose()
 
-	pass.Add(engine.TargetData())
 	pass.AddConstantPropagationPass()
 	pass.AddInstructionCombiningPass()
 	pass.AddPromoteMemoryToRegisterPass()