+++ llvm/docs/ReleaseNotes.rst
@@ -42,6 +42,8 @@ Non-comprehensive list of changes in this release
 
 * The C API function LLVMLinkModules has been removed.
 
+* The C API function LLVMAddTargetData has been removed.
+
 .. NOTE
    For small 1-3 sentence descriptions, just add an entry at the end of
    this list. If your description won't fit comfortably in one bullet